+++ tests/test_helpers_grpc_tracer.py
@@ -1,0 +1,5 @@
+from logtracer.helpers.grpc.tracer import GRPCTracer
+
+
+def test_GRPCTracer():
+    grpc_tracer = GRPCTracer()
